requirements.txt
Package       Version
------------- -------
aniso8601     8.0.0  
click         7.1.1  
DBUtils       1.3    
Flask         1.1.1  
Flask-Cors    3.0.8  
Flask-RESTful 0.3.8  
itsdangerous  1.1.0  
Jinja2        2.11.1 
MarkupSafe    1.1.1  
pip           19.0.3 
PyMySQL       0.9.3  
python-magic  0.4.15 
pytz          2019.3 
redis         3.4.1  
setuptools    40.8.0 
six           1.14.0 
SQLAlchemy    1.3.15 
Werkzeug      1.0.0  
